Move BarsAnimator to its own file and split private extension for UIScrollView so that each class can only use the necessary methods.
